Migrate project to SDL3 and modern SDL3 sub-libraries
- Update CMakeLists.txt to link against SDL3::SDL3, SDL3_image, SDL3_mixer, SDL3_ttf.
- Update GraphicsEngine to use SDL3 float-based rendering APIs (SDL_FRect, SDL_RenderRect).
- Update EventEngine to use SDL3 mouse state and event names.
- Update Timer to use Uint64 for SDL_GetTicks.
- Redesign AudioEngine and ResourceManager around the modern SDL3_mixer API utilizing MIX_Mixer, MIX_Track, and MIX_Audio objects rather than legacy Mix_Chunk or Mix_Music.
- Update XCube2d engine setup to modern SDL_Init checks and logical CPU functions.

+++ src/engine/Timer.cpp
@@ -10,10 +10,10 @@ void Timer::reset() {
 	time = 0;
 }
 
-Uint32 Timer::getTime() {
+Uint64 Timer::getTime() {
 	return time;
 }
 
-Uint32 Timer::getElapsed() {
+Uint64 Timer::getElapsed() {
 	return SDL_GetTicks() - time;
 }
